Add JavaDocs for canteen related code SLAP TextUi into different classes
