+++ src/main/java/seedu/navi/exceptions/HCAndMOCriteriaError.java
@@ -1,0 +1,4 @@
+package seedu.navi.exceptions;
+
+public class HCAndMOCriteriaError extends RuntimeException {
+}
