--- src/main/java/seedu/navi/exceptions/HCAndMOCrtieriaError.java
@@ -1,4 +1,0 @@
-package seedu.navi.exceptions;
-
-public class HCAndMOCrtieriaError extends RuntimeException {
-}
